requirements.txt
anyio==4.10.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.112.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pydantic==1.10.15
PyJWT==2.9.0
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.6
typing_extensions==4.14.1
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
